requirements.txt
streamlit
openai
python-dotenv
PyMuPDF
fitz
